backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Database and storage
pymongo==4.10.1
boto3==1.35.0

# HTTP and API clients
requests==2.32.3
httpx==0.27.2

# AI/ML integrations
openai==1.51.0
google-generativeai==0.8.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==4.2.0
email-validator==2.2.0

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Logging and utilities
loguru==0.7.2

# CORS support
fastapi-cors==0.0.6
